Check existence of impervious surface raster data

This function helps the user find the appropriate data from NLCD impervious surface raster data. It creates the necessary directory structure and provides guidance if the data is missing.

Usage

download_or_check_impervious(
  edition = "Annual_NLCD_FctImp_2024_CU_C1V1.tif",
  return_status = FALSE
)

Arguments

edition

Character string specifying the NLCD data edition to use. Defaults to "Annual_NLCD_FctImp_2024_CU_C1V1.tif".

return_status

Boolean where FALSE returns nothing, and TRUE returns the staus of the data. TRUE is returned if appropriate data is found.

Value

Function produces no output if ready to proceed; produces messaging if data is not available to the user.

Details

Data should be downloaded from the USGS NLCD. Data files are quite large (~1GB).

Next, the zip file should be unpacked and the .tif file should be placed in the urbanr_data directory inside the working directory.

Alternatively, you can download a different snapshot from the NLCD here and change the edition argument to that .tif filename. File should still be placed in the proper directory.

In the event that the data is no longer available from NLCD, the 2024 snapshot can also be found on Figshare.

Examples

download_or_check_impervious()
#> ✔ Data directory created at /home/runner/work/urbanr/urbanr/docs/reference
#> ✖ Data not found! Please first download the data from National Land Cover
#>   Database (NLCD). You can search for more snapshots here.
#> ℹ Next, add the .tif file to your working directory urbanr_data folder
#>   (/home/runner/work/urbanr/urbanr/docs/reference/urbanr_data).